Blocks are transported to storage server immediately
The upload is also carried out in the background to speed things up.
Immediate uploading means that the space necessary to create the backup
on local machine is at most configured BLOCK_SIZE plus or minus
compression and encryption, and the size of the index file.

If BLOCK_SIZE is 128K estimating the storage required for the index file
is easy - it is 1GB per every 1TB of backup.
If your blocks are 256K then it's 1GB per every 2TB, etc.
Basically, the bigger your blocks, the smaller your index.
However, smaller blocks yield better opportunities for deduplication.

All in all, you have to consider your circumstances and decide what
works best for you.

WARNING! Although mixing block sizes for different backups should in
theory work it is not recommended. For example, don't use 128KB blocks
at first and then change to 512KB blocks. I didn't research it but I
have a gut feeling that it's not a very good idea.
We already run the risk of damaging the backup due to the Dirichelet's
law (even though a pretty good hash function is used to identify the
blocks - SHA512) so do not mess with the scheme if you rely on your
backups.